Move all system tests into the api-tests package.
