+++ .changeset/47ddd0b6/changes.md
@@ -1,0 +1,1 @@
+- Add `matchFilter` and `runQuery` functions.
