+++ api-tests/README.md
@@ -1,0 +1,6 @@
+# API Tests
+
+Keystone, as a system, can be summarised as `schema => ({ api, adminUI })`.
+This directory contains system level test to ensure that `schema => ({ api })` behaves as expected.
+These system level tests should be considered complimentary to the unit tests contained in each package.
+Tests for `schema => ({ adminUI })` can be found in the `test-projects` directory.
