+++ api-tests/fields.test.js
@@ -1,39 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const cuid = require('cuid');
-const { graphqlRequest, multiAdapterRunners, setupServer } = require('@keystone-alpha/test-utils');
-
-const sorted = (arr, keyFn) => {
-  arr = [...arr];
-  arr.sort((a, b) => {
-    a = keyFn(a);
-    b = keyFn(b);
-    if (a < b) {
-      return -1;
-    }
-    if (a > b) {
-      return 1;
-    }
-    return 0;
-  });
-  return arr;
-};
-
-export const runQuery = (server, snippet) => {
-  return graphqlRequest({
-    server,
-    query: `query { ${snippet} }`,
-  }).then(res => res.body.data);
-};
-
-export const matchFilter = (server, gqlArgs, fields, target, sortkey) => {
-  gqlArgs = gqlArgs ? `(${gqlArgs})` : '';
-  const snippet = `allTests ${gqlArgs} ${fields}`;
-  return runQuery(server, snippet).then(data => {
-    const value = sortkey ? sorted(data.allTests || [], i => i[sortkey]) : data.allTests;
-    expect(value).toEqual(target);
-  });
-};
+const { multiAdapterRunners, setupServer } = require('@keystone-alpha/test-utils');
 
 // `mongodb-memory-server` downloads a binary on first run in CI, which can take
 // a while, so we bump up the timeout here.
