+++ api-tests/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@keystone-alpha/api-tests",
+  "private": true,
+  "version": "0.0.0",
+  "author": "The Keystone Development Team",
+  "license": "MIT",
+  "engines": {
+    "node": ">=8.4.0"
+  },
+  "devDependencies": {
+    "@keystone-alpha/adapter-mongoose": "^1.0.0",
+    "@keystone-alpha/fields": "^1.0.0",
+    "@keystone-alpha/keystone": "^1.0.0",
+    "@keystone-alpha/server": "^1.0.0",
+    "@keystone-alpha/test-utils": "^1.0.0",
+    "body-parser": "^1.18.2",
+    "cookie-signature": "^1.1.0",
+    "cuid": "^2.1.1",
+    "supertest-light": "^1.0.2",
+    "testcheck": "^1.0.0-rc.2"
+  }
+}
