+++ test-projects/access-control/package.json
@@ -26,8 +26,6 @@
     "@keystone-alpha/utils": "^1.0.0"
   },
   "devDependencies": {
-    "@keystone-alpha/test-utils": "^1.0.0",
-    "cuid": "^2.1.1",
     "cypress": "^3.1.5",
     "dotenv-safe": "^6.0.0",
     "execa": "1.0.0",
