+++ test-projects/basic/package.json
@@ -29,8 +29,6 @@
     "react": "^16.8.0"
   },
   "devDependencies": {
-    "@keystone-alpha/test-utils": "^1.0.0",
-    "cuid": "^2.1.1",
     "cypress": "^3.1.5",
     "dotenv-safe": "^6.0.0",
     "execa": "1.0.0",
@@ -38,7 +36,6 @@
     "mocha-junit-reporter": "^1.18.0",
     "mocha-multi-reporters": "^1.1.7",
     "pino-colada": "^1.4.4",
-    "start-server-and-test": "^1.7.11",
-    "testcheck": "^1.0.0-rc.2"
+    "start-server-and-test": "^1.7.11"
   }
 }
