+++ test-projects/login/package.json
@@ -26,9 +26,6 @@
     "body-parser": "^1.18.2"
   },
   "devDependencies": {
-    "@keystone-alpha/test-utils": "^1.0.0",
-    "cookie-signature": "^1.1.0",
-    "cuid": "^2.1.1",
     "cypress": "^3.1.5",
     "dotenv-safe": "^6.0.0",
     "execa": "1.0.0",
@@ -36,7 +33,6 @@
     "mocha-junit-reporter": "^1.18.0",
     "mocha-multi-reporters": "^1.1.7",
     "pino-colada": "^1.4.4",
-    "start-server-and-test": "^1.7.11",
-    "supertest-light": "^1.0.2"
+    "start-server-and-test": "^1.7.11"
   }
 }
